compiler warnings

+++ libunagipai/ata.c
@@ -32,7 +32,7 @@ static void ata_read_buffer(uint16_t* buffer) {
 
 static void ata_read_buffer_dma(uint16_t* buffer) {
 	dma_begin();
-	dma_transferblock_fromregister(0x800000, buffer, 256);
+	dma_transferblock_fromregister(0x800000, (uint32_t) buffer, 256);
 	dma_commit();
 }
